Disease Detection Based on Symptoms
Scope & Objective : 
Machine Learning applications in healthcare and biomedical domain has led to early disease detection and better diagnosis. This has enhanced patient care in recent times. Studies have shown that people take the help of the internet for any possible health-related issues. The problem with this approach is that the search engines provide bulk information in scattered format from which it is difficult to conclude.
The detection of disease based on disease is a complex game. Being unfamiliar with biological terms, the users feed the symptoms in non-technical or natural terms which add complexity in predicting diseases. The main objective is to develop a novel architecture that could accept and handle such type of user queries by employing techniques like query expansion using a thesaurus, synonym matching, and symptom suggestion that will allow disease prediction with greater accuracy based on user input. Data is scraped from the web and generated dataset can be used in future research. Query search retrieval and matching are used in such problems to achieve prediction. 

Business Problem Statement : 
The project proposes a model in which the user can input unstructured symptoms or select the symptoms suggested by the system, based on which, a list of probable diseases is provided back to the user. Further, the user can select any of the output diseases to get more information about its other symptoms, causes, diagnosis, possible treatment, etc. to help the user better understand the disease and current medical condition. The system also suggests other symptoms based on the ones that the user has input.

Data Sources : 
The dataset of disease and their symptoms has been scraped from the web by running the Python script. The dataset consists of diseases and their symptoms, which are fetched from the following sources
•	Diseases: The list of diseases has been retrieved from the National Health Portal of India (https://www.nhp.gov.in/disease-a-z), developed and maintained by Centre for Health Informatics (CHI)
•	Symptoms: The script uses the Google Search package to perform searching and fetch the disease's Wikipedia page among the various search results obtained. The HTML code of the page is processed to fetch the symptoms of the disease using the 'infobox' available on the Wikipedia page.
The scraping script fetches over 150+ different diseases that form the label and 500+ symptoms. The symptoms are then pre-processed to remove similar symptoms with different names.